Refactor CID store as plugin
