+++ modules/nextflow/src/main/groovy/nextflow/trace/DefaultObserverFactory.groovy
@@ -3,9 +3,6 @@ package nextflow.trace
 import java.nio.file.Path
 
 import nextflow.Session
-import nextflow.data.cid.CidObserver
-import nextflow.data.cid.CidStoreFactory
-
 /**
  * Creates Nextflow observes object
  * 
@@ -27,14 +24,7 @@ class DefaultObserverFactory implements TraceObserverFactory {
         createTimelineObserver(result)
         createDagObserver(result)
         createAnsiLogObserver(result)
-        createCidObserver(result)
         return result
-    }
-
-    protected void createCidObserver(Collection<TraceObserver> result) {
-        final store = CidStoreFactory.getOrCreate(session)
-        if( store )
-            result.add( new CidObserver(this.session, store) )
     }
 
     protected void createAnsiLogObserver(Collection<TraceObserver> result) {
