+++ modules/nextflow/src/main/resources/META-INF/build-info.properties
@@ -1,4 +1,4 @@
-build=5931
+build=5932
 version=25.01.0-edge
-timestamp=1739357731332
-commitId=e0916a4b8
+timestamp=1741733332268
+commitId=3e2ca194a
