+++ modules/nextflow/src/main/resources/META-INF/plugins-info.txt
@@ -1,5 +1,6 @@
 nf-amazon@2.12.0
 nf-azure@1.13.0
+nf-cid@1.0.0
 nf-cloudcache@0.4.3
 nf-codecommit@0.2.3
 nf-console@1.1.5
