+++ plugins/nf-cid/src/main/nextflow/data/cid/CidHistoryRecord.groovy
