+++ plugins/nf-cid/src/main/nextflow/data/cid/CidObserver.groovy
