+++ plugins/nf-cid/src/main/nextflow/data/cid/CidStore.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2024, Seqera Labs
+ * Copyright 2013-2025, Seqera Labs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@
 package nextflow.data.cid
 
 import java.nio.file.Path
-import java.util.function.Consumer
 
 import groovy.transform.CompileStatic
 import nextflow.data.config.DataConfig
