+++ plugins/nf-cid/src/main/nextflow/data/cid/CidStoreFactory.groovy
