+++ plugins/nf-cid/src/main/nextflow/data/cid/DefaultCidStore.groovy
