+++ plugins/nf-cid/src/main/nextflow/data/cid/fs/CidFileSystem.groovy
