+++ plugins/nf-cid/src/main/nextflow/data/cid/fs/CidFileSystemProvider.groovy
