+++ plugins/nf-cid/src/main/nextflow/data/cid/fs/CidPathFactory.groovy
