+++ plugins/nf-cid/src/main/nextflow/data/cid/model/DataPath.groovy
