+++ plugins/nf-cid/src/main/nextflow/data/cid/model/Parameter.groovy
