+++ plugins/nf-cid/src/main/nextflow/data/cid/model/Workflow.groovy
