+++ plugins/nf-cid/src/resources/META-INF/extensions.idx
@@ -1,5 +1,5 @@
 #
-# Copyright 2013-2024, Seqera Labs
+# Copyright 2013-2025, Seqera Labs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,15 +14,6 @@
 # limitations under the License.
 #
 
-nextflow.trace.DefaultObserverFactory
-nextflow.util.DefaultSerializers
-nextflow.secret.LocalSecretsProvider
-nextflow.cache.DefaultCacheFactory
-nextflow.scm.RepositoryFactory
-nextflow.container.resolver.DefaultContainerResolver
-nextflow.mail.SendMailProvider
-nextflow.mail.SimpleMailProvider
-nextflow.mail.JavaMailProvider
-nextflow.processor.tip.DefaultTaskTipProvider
-nextflow.fusion.FusionTokenDefault
-
+nextflow.data.cid.CidStoreFactory
+nextflow.data.cid.CidObserverFactory
+nextflow.data.cid.operation.CidOperationImpl
