+++ plugins/nf-cid/src/test/nextflow/data/cid/CidObserverTest.groovy
