+++ plugins/nf-cid/src/test/nextflow/data/cid/DefaultCidStoreTest.groovy
