+++ plugins/nf-cid/src/test/nextflow/data/cid/fs/CidFileSystemProviderTest.groovy
