+++ plugins/nf-cid/src/test/nextflow/data/cid/fs/CidPathTest.groovy
