+++ plugins/nf-cid/src/test/nextflow/data/cid/fs/CifPathFactoryTest.groovy
