+++ plugins/nf-cid/src/test/nextflow/data/config/DataConfigTest.groovy
