+++ settings.gradle
@@ -40,3 +40,4 @@ include 'plugins:nf-azure'
 include 'plugins:nf-codecommit'
 include 'plugins:nf-wave'
 include 'plugins:nf-cloudcache'
+include 'plugins:nf-cid'
